Add files via upload
My first project that I have ever done! First time using HTML/CSS to code and make my own webbsite.  I decided to do a Yoga webbsite that is very simple to navigate and get around. 
I have made a Home, schedule, pricing and contact page and added some information about the benefits about yoga in the home page with some pictures of different poses that one might do in yoga. 
There is a navigation bar that is very easily spotted is on top of the page with four different option to navigate to the rest of the webbsite with. 
Ive added in the schedule page on what someone might expect in a whole week doing yoga at this specific place. Ive added time stamps that are fixed and it spans from monday to sunday with different activities during the week. I have also added a google maps location on the schedule page to show people where  its located and an address to the studio so people know where to find it. 
Ive added on the pricing page 3 different membership option aswell and what comes with them. 
And on the contact page ive added a form so you fill in and contact the staff to ask for more information or simply state what you would like to achive with doing yoga. 
Its very easy to navigate as said in the beginning and a very simple design to it all.